+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Normalized all [alpha] values to be on a `0` to `255` scale to be in line
 with Flutter's [Color] class. They were on a `0` to `1` scale.
 
+* Renamed [HsvColor] to [HsbColor] to avoid a naming conflict with Flutter's
+[Color] class's [value] parameter.
+
 ## [0.2.8] - March 29, 2020
 
 * Removed [isBlack] and [isWhite] conversion condition from the CMYK color space.